requirements.txt
streamlit
pandas
openpyxl
Pillow
pyautogui
pywinauto
odfpy
xlsxwriter
opencv-python


